Support both versions 1 & 2 of the AWS SDK for Java
Jon Norman mentioned that he was looking to use `play-secret-rotation` on
a new project that was using version **2** of the AWS SDK for Java
(released in November 2018):

https://docs.aws.amazon.com/sdk-for-java/v2/developer-guide/welcome.html
https://aws.amazon.com/blogs/developer/aws-sdk-for-java-2-x-released/

This commit creates two new artifacts for `play-secret-rotation`:

* `aws-parameterstore-sdk-v1`
* `aws-parameterstore-sdk-v2`

Consumers of the `play-secret-rotation` can decide which artifact to
import, and then setup their `SecretSupplier` appropriately with
`com.gu.play.secretrotation.aws.parameterstore.AwsSdkV1` or
`com.gu.play.secretrotation.aws.parameterstore.AwsSdkV2`, eg:

```scala
import com.gu.play.secretrotation._

val secretStateSupplier: SnapshotProvider = {
  import com.gu.play.secretrotation.aws.parameterstore

  new parameterstore.SecretSupplier(
    transitionTiming, "/Example/PlayAppSecret",
    parameterstore.AwsSdkV1(AWSSimpleSystemsManagementClientBuilder.defaultClient())
  )
}
```

Happily for the development of the `play-secret-rotation` project, this
arrangement doesn't require us to keep two separate Git branches in sync -
the support for versions 1 & 2 of the AWS SDK for Java can co-exist in the
same project file tree - just compiled as distinct sub-projects under
the root sbt project.

Another small change, which has been fortunate, is that both versions 1 & 2
of the AWS SDK for ParameterStore now support *reading the metadata* when
reading multiple parameter values. This means we no longer need
`describeParameters` and it's associated permission in cloudformation.

There's a branch in Ophan which adopts this version of
`play-secret-rotation` and shows some of the tweaks that have taken place
with this update:

https://github.com/guardian/ophan/commits/update-play-secret-rotation-to-test-aws-sdk-versions-support
