+++ aws-parameterstore/secret-supplier/aws-sdk-v2/AwsSdkV2.scala
@@ -1,0 +1,11 @@
+package com.gu.play.secretrotation.aws.parameterstore
+
+import software.amazon.awssdk.services.ssm.SsmClient
+
+import scala.collection.JavaConverters._
+
+case class AwsSdkV2(ssmClient: SsmClient) extends MinimalAwsSdkWrapper{
+  override def fetchValues(parameters: Seq[String]): Seq[ParameterValue] = ssmClient.getParameters(
+    _.withDecryption(true).names(parameters.asJavaCollection)
+  ).parameters.asScala.map(p => ParameterValue(p.value, Metadata(p.version, p.lastModifiedDate)))
+}
