+++ build.sbt
@@ -17,28 +17,43 @@ lazy val core =
     )
   )
 
-val awsSdkVersion = "1.11.313"
+lazy val `aws-parameterstore-secret-supplier-base` =
+  project.in(file("aws-parameterstore/secret-supplier")).settings(baseSettings: _*).dependsOn(core)
 
-val awsSsm = "com.amazonaws" % "aws-java-sdk-ssm" % awsSdkVersion
-
-lazy val `aws-parameterstore` = project.in(file("aws-parameterstore/state-supplier")).settings(baseSettings: _*).dependsOn(core).settings(
-  libraryDependencies += awsSsm
+val awsSdkForVersion = Map(
+  1 -> "com.amazonaws" % "aws-java-sdk-ssm" % "1.11.524",
+  2 -> "software.amazon.awssdk" % "ssm" % "2.5.15"
 )
+
+def awsParameterStoreWithSdkVersion(version: Int)=
+  Project(s"aws-parameterstore-sdk-v$version", file(s"aws-parameterstore/secret-supplier/aws-sdk-v$version"))
+  .settings(baseSettings: _*)
+  .dependsOn(`aws-parameterstore-secret-supplier-base`)
+  .settings(libraryDependencies += awsSdkForVersion(version))
+
+lazy val `aws-parameterstore-sdk-v1` = awsParameterStoreWithSdkVersion(1)
+lazy val `aws-parameterstore-sdk-v2` = awsParameterStoreWithSdkVersion(2)
 
 lazy val `aws-parameterstore-lambda` = project.in(file("aws-parameterstore/lambda"))
   .settings(baseSettings: _*).dependsOn(`secret-generator`).settings(
   libraryDependencies ++= Seq(
     "com.amazonaws" % "aws-lambda-java-core" % "1.2.0",
     "com.amazonaws" % "aws-lambda-java-events" % "2.0.2",
-    awsSsm
+    awsSdkForVersion(1)
   )
 )
 
 lazy val `secret-generator` = project.settings(baseSettings: _*)
 
 lazy val `play-secret-rotation-root` = (project in file("."))
-  .aggregate(core, `aws-parameterstore`, `aws-parameterstore-lambda`).
-  settings(baseSettings: _*).settings(
+  .aggregate(
+    core,
+    `aws-parameterstore-secret-supplier-base`,
+    `aws-parameterstore-sdk-v1`,
+    `aws-parameterstore-sdk-v2`,
+    `aws-parameterstore-lambda`
+  )
+  .settings(baseSettings: _*).settings(
   publishArtifact := false,
   publish := {},
   publishLocal := {},
